# Conflicts:
#	parallel-consumer-core/pom.xml
#	parallel-consumer-examples/parallel-consumer-example-core/pom.xml
#	parallel-consumer-examples/parallel-consumer-example-reactor/pom.xml
#	parallel-consumer-examples/parallel-consumer-example-streams/pom.xml
#	parallel-consumer-examples/parallel-consumer-example-vertx/pom.xml
#	parallel-consumer-examples/pom.xml
#	parallel-consumer-reactor/pom.xml
#	parallel-consumer-vertx/pom.xml
#	pom.xml

+++ parallel-consumer-core/src/test-integration/java/io/confluent/parallelconsumer/integrationTests/CloseAndOpenOffsetTest.java
@@ -223,7 +223,7 @@ class CloseAndOpenOffsetTest extends BrokerIntegrationTest<String, String> {
         log.debug("Sending {} messages to {}", quantity, topic);
         var futures = new ArrayList<Future<RecordMetadata>>();
         // async
-        for (Integer index : Range.range(quantity)) {
+        for (Long index : Range.range(quantity)) {
             Future<RecordMetadata> send = getKcu().getProducer().send(new ProducerRecord<>(topic, partition, index.toString(), index.toString()));
             futures.add(send);
         }
